Refactor D-FINE module by removing unused functions and integrating dfine_bbox2distance in DFINECriterion
