+++ src/otx/algo/detection/losses/dfine_loss.py
@@ -16,113 +16,7 @@ from torchvision.ops import box_convert
 
 from otx.algo.common.utils.assigners.hungarian_matcher import HungarianMatcher
 from otx.algo.common.utils.bbox_overlaps import bbox_overlaps
-from otx.algo.detection.heads.dfine_decoder import weighting_function
-
-
-def translate_gt(gt: Tensor, reg_max: int, reg_scale: Tensor, up: Tensor) -> tuple[Tensor, Tensor, Tensor]:
-    """Decodes bounding box ground truth (GT) values into distribution-based GT representations.
-
-    This function maps continuous GT values into discrete distribution bins, which can be used
-    for regression tasks in object detection models.
-
-    It calculates the indices of the closest bins to each GT value and assigns interpolation weights
-    to these bins based on their proximity to the GT value.
-
-    In the paper:
-        'a' (up) controlling the upper bounds.
-        'c' (reg_scale) controlling the curvature.
-
-    Args:
-        gt (Tensor): Ground truth bounding box values, shape (N, ).
-        reg_max (int): Maximum number of discrete bins for the distribution.
-        reg_scale (Tensor): Controls the curvature of the Weighting Function.
-        up (Tensor): Controls the upper bounds of the Weighting Function.
-
-    Returns:
-        Tuple[Tensor, Tensor, Tensor]:
-            - indices (Tensor): Index of the left bin closest to each GT value, shape (N, ).
-            - weight_right (Tensor): Weight assigned to the right bin, shape (N, ).
-            - weight_left (Tensor): Weight assigned to the left bin, shape (N, ).
-    """
-    gt = gt.reshape(-1)
-    function_values = weighting_function(reg_max, up, reg_scale)
-
-    # Find the closest left-side indices for each value
-    diffs = function_values.unsqueeze(0) - gt.unsqueeze(1)
-    mask = diffs <= 0
-    closest_left_indices = torch.sum(mask, dim=1) - 1
-
-    # Calculate the weights for the interpolation
-    indices = closest_left_indices.float()
-
-    weight_right = torch.zeros_like(indices)
-    weight_left = torch.zeros_like(indices)
-
-    valid_idx_mask = (indices >= 0) & (indices < reg_max)
-    valid_indices = indices[valid_idx_mask].long()
-
-    # Obtain distances
-    left_values = function_values[valid_indices]
-    right_values = function_values[valid_indices + 1]
-
-    left_diffs = torch.abs(gt[valid_idx_mask] - left_values)
-    right_diffs = torch.abs(right_values - gt[valid_idx_mask])
-
-    # Valid weights
-    weight_right[valid_idx_mask] = left_diffs / (left_diffs + right_diffs)
-    weight_left[valid_idx_mask] = 1.0 - weight_right[valid_idx_mask]
-
-    # Invalid weights (out of range)
-    invalid_idx_mask_neg = indices < 0
-    weight_right[invalid_idx_mask_neg] = 0.0
-    weight_left[invalid_idx_mask_neg] = 1.0
-    indices[invalid_idx_mask_neg] = 0.0
-
-    invalid_idx_mask_pos = indices >= reg_max
-    weight_right[invalid_idx_mask_pos] = 1.0
-    weight_left[invalid_idx_mask_pos] = 0.0
-    indices[invalid_idx_mask_pos] = reg_max - 0.1
-
-    return indices, weight_right, weight_left
-
-
-def bbox2distance(
-    points: Tensor,
-    bbox: Tensor,
-    reg_max: int,
-    reg_scale: Tensor,
-    up: Tensor,
-    eps: float = 0.1,
-) -> tuple[Tensor, Tensor, Tensor]:
-    """Converts bounding box coordinates to distances from a reference point.
-
-    Refer to D-Fine: https://github.com/Peterande/D-FINE.
-
-    Args:
-        points (Tensor): (n, 4) [x, y, w, h], where (x, y) is the center.
-        bbox (Tensor): (n, 4) bounding boxes in "xyxy" format.
-        reg_max (float): Maximum bin value.
-        reg_scale (float): Controling curvarture of W(n).
-        up (Tensor): Controling upper bounds of W(n).
-        eps (float): Small value to ensure target < reg_max.
-
-    Returns:
-        Tuple[Tensor, Tensor, Tensor]:
-            - indices (Tensor): Index of the left bin closest to each GT value, shape (N, ).
-            - weight_right (Tensor): Weight assigned to the right bin, shape (N, ).
-            - weight_left (Tensor): Weight assigned to the left bin, shape (N, ).
-    """
-    reg_scale = abs(reg_scale)
-    # ϕ = (dᴳᵀ- d⁰) / {H, H, W, W}
-    left = (points[:, 0] - bbox[:, 0]) / (points[..., 2] / reg_scale + 1e-16) - 0.5 * reg_scale
-    top = (points[:, 1] - bbox[:, 1]) / (points[..., 3] / reg_scale + 1e-16) - 0.5 * reg_scale
-    right = (bbox[:, 2] - points[:, 0]) / (points[..., 2] / reg_scale + 1e-16) - 0.5 * reg_scale
-    bottom = (bbox[:, 3] - points[:, 1]) / (points[..., 3] / reg_scale + 1e-16) - 0.5 * reg_scale
-    four_lens = torch.stack([left, top, right, bottom], -1)
-    four_lens, weight_right, weight_left = translate_gt(four_lens, reg_max, reg_scale, up)
-    if reg_max is not None:
-        four_lens = four_lens.clamp(min=0, max=reg_max - eps)
-    return four_lens.reshape(-1).detach(), weight_right.detach(), weight_left.detach()
+from otx.algo.detection.utils.utils import dfine_bbox2distance
 
 
 class DFINECriterion(nn.Module):
@@ -279,7 +173,7 @@ class DFINECriterion(nn.Module):
             pred_corners = outputs["pred_corners"][idx].reshape(-1, (self.reg_max + 1))
             ref_points = outputs["ref_points"][idx].detach()
             with torch.no_grad():
-                target_corners, weight_right, weight_left = bbox2distance(
+                target_corners, weight_right, weight_left = dfine_bbox2distance(
                     ref_points,
                     box_convert(target_boxes, in_fmt="cxcywh", out_fmt="xyxy"),
                     self.reg_max,
